+++ loom/src/loom/ops/successor_verify.h
@@ -22,7 +22,7 @@ extern "C" {
 // for count mismatches and STRUCTURE/026 for type mismatches.
 iree_status_t loom_ops_verify_successor_args(
     const loom_module_t* module, iree_diagnostic_emitter_t emitter,
-    const loom_op_t* op, iree_string_view_t op_name, uint8_t successor_index,
+    const loom_op_t* op, iree_string_view_t op_name, uint16_t successor_index,
     const loom_block_t* target, const loom_value_id_t* args,
     iree_host_size_t arg_count);
